Projects/converters/91(pdf_text_to_speech)/requirements.txt
pdfquery==0.4.3
google-cloud-texttospeech==2.14.1